Corrected the I2S_SDIN and I2S_SDOUT pins
Previously, the I2S_SDIN and I2S_SDOUT pins were connected to pins 7 and
8 of the Teensy package. However, they should have been connected to
Arduino digital I/O pins 7 and 8 instead. These pins are actually pins
10 and 9 of the Teensy package. Pins 10 and 9 of the Teensy package were
instead connected to T_CS and D_LEDEN.

In order to fix this problem, the signals on pins 7 and 10 (I2S_SDIN and
T_CS) were swapped. Also the signals on pins 8 and 9 (I2S_SDOUT and
D_LEDEN) were swapped. This was done with cuts and jumper wires on
PCB2001A. It will be fixed in etch on PCB2001B.