Fixed LED staying on when it shouldn't be
Previously there were two times when the strobe LED stayed on when it
was supposed to be off. First, when the system started, the PWM pin
would float and would not turn off the LED. Secondly, the non-inverting
input of the op amp would not go low enough to completely turn off off
the LED.

The startup problem was fixed by adding a 10k pull-up resistor to the
PWM signal so that it would be high during the time before the code
initialized that signal to be an output and actively drives it.

The second problem was fixed by adding a 22k pull-up to 5V on the
inverting input of the op amp. This kept the inverting input at a higher
voltage and allowed the non-inverting input to completely turn the op
amp output off.