Increased voltage from 36V to 48V
Previously, the op amp was supplied voltage directly from the 36V power
supply. However, this was not high enough to drive the COB LED to high
currents. Now, the voltage has been increased to 48V, but this
necessitated regulating the voltage to the MC34071A because it cannot
handle 48V.