Changed current-sense resistor to allow higher current
The current-sense resistor R43 was changed from 0.5-ohm to 0.05-ohm to
allow more current while keeping the voltage across the resistor lower.
Additionally, a pull-down resistor was added to the op amp non-inverting
input so that a current-sense voltage of 0V to 12V gets converted to a
range of 0.2V to 4.9V at the non-inverting input of the op amp.

Additionally, the voltage divider on the Vref input of the MCP4706A DAC
was removed because the VDD power input will be used as the reference
instead.